Refactor P/T model structures and enhance error handling
- Introduced new `bug` and `pt` modules in `src/lib.rs` for better organization of P/T model features.
- Updated `InputArc` and `OutputArc` structs to use `u64` for the `weight` field, allowing for larger token counts.
- Enhanced `RuntimeError` with new error types for invalid P/T models and capacity issues.
- Modified `PtMarking` to support `u64` tokens and added new methods for better token management.
- Updated tests to reflect changes in weight handling and token representation.

+++ src/core/model.rs
@@ -50,14 +50,14 @@ pub struct TransitionDecl {
 pub struct InputArc {
     pub place: PlaceId,
     pub pattern: Pattern,
-    pub weight: u32,
+    pub weight: u64,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct OutputArc {
     pub place: PlaceId,
     pub term: Term,
-    pub weight: u32,
+    pub weight: u64,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
